In the run-directory Makefile, pass APM variable to compilation commands
In the previous commits, we had defined an APM variable in the run
directory Makefile template: runs/shared_inputs/Makefiles/Makefile.

We had forgotten to actually pass this variable to the compilation
commands further down in the Makefile.  This is now fixed.

We have also updated the perl/difftest/Makefile accordingly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>