Update plot directory and scripts in benchmark run directory
The "output" directory has been renamed to "Plots" to make it more clear
that the directory will contain benchmark plots, not GEOS-Chem output.
The scripts within this directory have also been cleaned up. We now utilize
GCPy for most of the benchmarking plots, but we still use GAMAP for the
budget output (because it is required for the metrics.pl script). The
plot_1mon.pro script has been updated to only create the GAMAP budget output
and not any of the other GAMAP benchmark plots.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>