Now put GEOS-Chem and HEMCO output files in OutputDir of run directory
To be more consistent with the run directory structure used in GCHP, we now
send output from History and HEMCO to the OutputDir subdirectory. The
OutputDir is created whenever running a unit test, copying a run directory,
or creating a difference test directory. OutputDir and its contents will be
removed when cleanFiles or cleanRunDirs is executed. The validate.pl script
has also been updated to check the contents of OutputDir in addition to the
bpch and netCDF files in the run directory.

The exception is that the GEOS-Chem restart files will still be saved to
the top-level run directory and not to OutputDir. This is done by specifying
the exact location and filename via the Restart.filename attribute in
HISTORY.rc and removing the Restart.template attribute.

Also fixed a bug in getRunInfo. The ENDTIME returned was incorrect because
the start index for hhmm in YYYYMMDDhhmm was off by 1.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>