Add O3 vertical mass flux diagnostic to fullchem HISTORY.rc files
The AdvFluxVert collection was already added to the code in an earlier
version. Here we add that collection to the HISTORY.rc.template files
for the full-chemistry mechanisms. It is only on by default for
benchmark simulations, and only for O3. This diagnostic is used to compute
the O3 STE in the benchmark output.

Also put diagnostics in alphabetical order, except for Restart and SpeciesConc
(top of list) which should be turned on in more cases and Boundary conditions
(bottom of list) which should be turned off unless proceeding to run nested
grid simulations.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>