Update gcUnitTest script to only compile once for single processor and once for multi-processor
The perl/gcUnitTest script now handles the single/multi-processor builds
and runs instead of the run directory Makefile which now simply handles
cleaning, building, and running GEOS-Chem.

As part of this update, the run directory Makefile is copied to the root
run directory and GEOS-Chem is only compiled once for sp (to geos.sp)
and once for mp (to geos.mp). Recompilation is not required when changing
met field or resolution. Additional compilations are, however, required for
simulations like APM, RRTMG, and TOMAS and those builds will be saved to
separate executable files. Compilation logs will be saved to the logs
directory and lastbuild files will remain in the root run directory with
the respective geos executable files. The cleanFiles script has been
updated to account for these new files.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>