Update gcUnitTest script to pass run directory to make command
The run directory is needed when compiling so that the getRunInfo script
can access the input.geos file to determine the simulation name and
start/end dates. The Makefile assumed that the current directory was the
run directory, but that is not the case when unit testing. This fix ensures
the proper run options (e.g. APM, TOMAS, RRTMG) are turned on when
running specialty simulations.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>