Turn on Yuan-processed MODIS LAI by default
A fix has been added to HEMCO to address the interpolation issue when reading
the Yuan-processed MODIS LAI product. We will now use this dataset as the
default.

Christoph Keller wrote:

   My suspicion is that the problem is with the ‘ExitSearch’ flag, set in
   GET_TIMEIDX on line 1506 (hcoio_read_std_mod.F90). In short, this flag
   scans for ‘discontinuous’ data sets. For example, the RCP scenarios
   contain monthly emissions for a few selected years (2005, 2010, 2020,
   2050, etc.). Just doing a linear interpolation between adjacent time
   slices doesn’t necessarily make sense for e.g. July 2007 (interpolation
   should be between July 2005 and July 2010, instead of December 2005 to
   January 2010). So the code attempts to identify this and then do a
   ‘smart’ interpolation by picking the correct monthly emissions for the
   lower and upper bound years. I suspect this is what is happening here,
   too.

To avoid issues with datasets like Yuan-processed MODIA LAI, we have added
a new cycle option to denote when datasets are discontinuous. Without the
"D" option appended to the "I" option, the continued search in Get_TimeIDX
will not be executed. The only discontinuous dataset currently supported in
GEOS-Chem is RCP, which has data for yeards 2005, 2010, 2020, ..., 2100.

Also fixed the years for a few datasets in HEMCO_Config.rc (GMI, ALD2 seawater,
MENO3, ETNO3) to match the years in the datafiles. Otherwise, with the
latest update, HEMCO will crash because it can't find data for the
preferred date.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>