Fix dates in several HEMCO_Config.rc files to reflect dates in data files
With updates made in 12.7.0, HEMCO is more strict when comparing dates listed
in HEMCO_Config.rc to dates available in the data files. Here we fix several
incorrect dates in HEMCO_Config.rc files to allow files to be read in properly.

Also consolidated Hg and tagHg shared_inputs files because there is currently no
difference except in the README files.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>