Make input.geos.template files in runs/2x25_*  pointers to files in runs/4x5_*
With FlexGrid, there are now no differences between 2x2.5 and 4x5 input.geos.template
files. The Grid Menu is uses tokens and is filled in by settings specified
in the unit tester input files. To reduce the number of input.geos.template
files that need to be updated, we have removed the 2x2.5 files and now simply
point to the 4x5 files. This will also simplify the impending migration of
run directories to the GEOS-Chem code repository.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>