Update boundary condition entries in HEMCO_Config.rc to read entire day at once
The hour setting for the BCs in HEMCO_Config.rc was previously 0-23, which
would tell HEMCO to update these fields every hour. This would cause an issue
after hour 21, for which there are no BC fields for the remainder of the day.
To avoid this error, we now read the BCs for all timestamps on a given day
at hour 0 and in hcoi_gc_main_mod.F we use new functions to determine if it's
time to update the boundary conditions. By default, this is done every 3
hours.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>