Add HEMCO_Config.template.RRTMG file in runs/shared_inputs/fullchem
This update fixes an issue that was reported as:
   https://github.com/geoschem/geos-chem/issues/235

Basically, the RRTMG rundirs linked to HEMCO_Config.template.tropchem,
which was configured to not read in the necessary initialization files
for RRTMG (i.e. RRTMG was set to "false" in the MASTER SWITCHES SECTION).

This is now fixed.  The RRTMG run directories now point to a
HEMCO_Config.template.RRTMG file with the proper settings.

This fix is slated for GEOS-Chem 12.7.2.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>