Alphabetize species in Advected Species Menu of input.geos for full-chemistry simulations
With the increasing number of species in the full-chemistry simulations, it's
becoming increasingly difficult to compare the list of species when they are
updated. To simplify things, we will list the species in alphabetical order in
the Advected Species Menu of input.geos. This also allows for easier comparison
with the species database, netCDF output files, etc. where species are listed
in alphabetical order as well.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>