Move APM and TOMAS species to end of advected species lists
There is code in aerosol_mod.F90 that maps hygroscopic aerosol species and
needs SO4, BCPI, OCPI, SALA, SALC to be the first five species in the
list to work properly. For now, to get around this the APM and TOMAS species
have been moved back to the end of the advected species list in input.geos
instead of being in alphabetical order with all of the advected species.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>