Include Makefile target 'build' in the phony list
This update allows using the 'build' target in the GNU makefile when
there is a subdirectory called 'build' in the run directory. If not
declared as phony then the build cannot be executed, exiting with the
following message:
   make: `build' is up to date.

NOTE: GNU make will be retired in GEOS-chem 13.0.0 at which point the
Makefile will be deleted and this will not longer be a potential issue.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>