Fix bug causing run fail if not starting simulation at YYYYMM01 00:00:00
This issue was caused by use of the Exact flag for NOAA_GMD_CH (surface
CH4) used in full chemistry simulations. The fix, which is temporary
pending further discussion of what the default behavior for surface CH4
should be, is to use the Range flag instead. This will cause HEMCO to use
the nearest timestamp of the dataset regardless of exact simulation
datetime.

To use the CMIP6 surface CH4 instead of NOAA_GMD then NOAA_GMD_CH4 should
be set to False in HEMCO_Config.rc, and CMIP6_Sfc_CH4 set to True. It is
currently not possible to switch between the two datasets in a continuous
run. The ability to do this AND prevent the bug that motivated this fix
will be in a future update if it is decided this is the target default
behavior. Please note that CMIP6 still uses the Exact flag so simulations
using CMIP6 (valid only for runs starting before 1979) must start at
midnight on the first day of the month. This also may change in a future
update.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>