Update perl scripts to copy species_database.yml file to run directories automatically
By default, the species_database.yml file should be located in the run directory
as is now specified in the input.geos file. The perl scripts for copying a
run directory, creating a difference test, and running a unit test have
now been updated to automatically copy that file from its current location
in CodeDir/Headers/ to the run directory. The cleanup scripts will remove
that file from the unit tester run directories.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>